refactor(fleet): extract status projection
